refactor(ui): extract Interfaces tab to ui/tabs/interfaces.rs
